move all internals into private module to discourage non-apiver compliant imports
