+++ django_business_metrics/_internal/__init__.py
@@ -1,0 +1,5 @@
+"""
+Internals of django-business-metrics implementation.
+
+This project uses ApiVer, and as such, all imports should be done from v* submodules.
+"""
