+++ django_business_metrics/v0/__init__.py
@@ -1,2 +1,2 @@
-from django_business_metrics.metrics import active_users, users
-from django_business_metrics.metrics_manager import BusinessMetricsManager
+from django_business_metrics._internal.metrics import active_users, users
+from django_business_metrics._internal.metrics_manager import BusinessMetricsManager
